Rename folder motor to apply, removed duplicate component for contact details, arrange order of forms
