+++ client/containers/apply/hair-color-form-container.jsx
@@ -2,22 +2,23 @@
 
 import React from 'react';
 
-import { updateSex }          from '../../actions/index';
-import Form                   from '../../presentations/motor/sex-form.jsx';
+import { updateHairColor }    from "../../actions/index";
+import Form                   from '../../presentations/apply/hair-color-form.jsx';
 import connectForm            from '../../helpers/connect-form';
 import navigateOnSubmit       from '../../helpers/navigate-on-submit';
 import * as dataPresent       from '../../helpers/data-present';
 
 const ConnectedForm = (props) => {
-  let continueDisabled  = !dataPresent.value(props.sex);
-  let onSubmit          = navigateOnSubmit('/about-me/appearance/eye', props);
-  let pageTitle         = 'About me: Sex identification';
+  let continueDisabled  = !dataPresent.value(props.hairColor);
+  let onSubmit          = navigateOnSubmit('/about-me/height', props);
+  let pageTitle         = 'About me: Hair color';
+
   return (
     <Form
       pageTitle         = { pageTitle }
       onSubmit          = { onSubmit }
       onChange          = { props.onChange }
-      selectedValue     = { props.sex }
+      selectedValue     = { props.hairColor }
       continueDisabled  = { continueDisabled }
     />
   );
@@ -25,8 +26,9 @@ const ConnectedForm = (props) => {
 
 function mapStateToProps(state) {
   return {
-    sex: state.application.sex
+    hairColor: state.application.hairColor
   };
 }
 
-export default connectForm(mapStateToProps, updateSex, ConnectedForm);
+export default connectForm(mapStateToProps, updateHairColor, ConnectedForm);
+
