+++ client/containers/apply/home-address-form-container.jsx
@@ -3,7 +3,7 @@
 import React from 'react';
 
 import { updateHomeAddress }   from "../../actions/index";
-import Form                    from "../../presentations/motor/home-address-form.jsx";
+import Form                    from "../../presentations/apply/home-address-form.jsx";
 import connectForm             from '../../helpers/connect-form';
 import navigateOnSubmit        from '../../helpers/navigate-on-submit';
 import * as dataPresent        from '../../helpers/data-present';
