+++ client/containers/apply/interstitial-address-container.jsx
@@ -5,7 +5,7 @@ import React                        from 'react';
 
 import alicePath                    from '../../helpers/alice-path';
 import { updateMailingAddress }     from "../../actions/index";
-import Form                         from "../../presentations/motor/interstitial-address-form.jsx";
+import Form                         from "../../presentations/apply/interstitial-address-form.jsx";
 import connectForm                  from '../../helpers/connect-form-interstitial';
 import navigateOnSubmit             from '../../helpers/navigate-on-submit';
 import * as dataPresent             from '../../helpers/data-present';
