+++ client/containers/apply/sex-selector-form-container.jsx
@@ -2,23 +2,22 @@
 
 import React from 'react';
 
-import { updateDateOfBirth }  from "../../actions/index";
-import Form                   from "../../presentations/motor/date-of-birth-form.jsx";
+import { updateSex }          from '../../actions/index';
+import Form                   from '../../presentations/apply/sex-form.jsx';
 import connectForm            from '../../helpers/connect-form';
 import navigateOnSubmit       from '../../helpers/navigate-on-submit';
 import * as dataPresent       from '../../helpers/data-present';
 
 const ConnectedForm = (props) => {
-  let onSubmit          =   navigateOnSubmit('/about-me/home-address', props);
-  let continueDisabled  =   !(dataPresent.date(props.dateOfBirth));
-  let pageTitle         =   'About me: Date of birth';
-
+  let continueDisabled  = !dataPresent.value(props.sex);
+  let onSubmit          = navigateOnSubmit('/about-me/appearance/eye', props);
+  let pageTitle         = 'About me: Sex identification';
   return (
     <Form
       pageTitle         = { pageTitle }
       onSubmit          = { onSubmit }
       onChange          = { props.onChange }
-      dateOfBirth       = { props.dateOfBirth }
+      selectedValue     = { props.sex }
       continueDisabled  = { continueDisabled }
     />
   );
@@ -26,8 +25,8 @@ const ConnectedForm = (props) => {
 
 function mapStateToProps(state) {
   return {
-    dateOfBirth: state.application.dateOfBirth
+    sex: state.application.sex
   };
 }
 
-export default connectForm(mapStateToProps, updateDateOfBirth, ConnectedForm);
+export default connectForm(mapStateToProps, updateSex, ConnectedForm);
