+++ client/presentations/home.jsx
@@ -14,6 +14,7 @@ const LinkListItem = (props) => {
 
 const linkData = [
   {description: 'summary',                    path: '/summary'},
+
   {description: 'legal name',                 path: '/about-me/legal-name'},
   {description: 'date of birth',              path: '/about-me/date-of-birth'},
   {description: 'home address',               path: '/about-me/home-address'},
@@ -24,17 +25,19 @@ const linkData = [
   {description: 'height',                     path: '/about-me/height'},
   {description: 'weight',                     path: '/about-me/weight'},
   {description: 'social security',            path: '/about-me/social-security'},
-  {description: 'contact details',            path: '/about-me/voter/email-phone'},
+
   {description: 'citizen status',             path: '/about-me/voter/am-citizen'},
-  {description: 'ballot by mail',             path: '/about-me/voter/ballot-by-mail'},
   {description: 'eligibility requirements',   path: '/about-me/voter/eligibility-requirements'},
+  {description: 'opt out',                    path: '/about-me/voter/opt-out'},
   {description: 'voter preferences intro',    path: '/about-me/voter/voter-preferences-intro'},
-  {description: 'contact methods',            path: '/about-me/voter/contact-choice'},
   {description: 'political party choose',     path: '/about-me/voter/political-party-choose'},
   {description: 'political party',            path: '/about-me/voter/political-party'},
   {description: 'ballot language',            path: '/about-me/voter/ballot-language'},
-  {description: 'success visit',              path: '/about-me/success-visit'},
-  {description: 'opt out',                    path: '/about-me/voter/opt-out'}
+  {description: 'ballot by mail',             path: '/about-me/voter/ballot-by-mail'},
+  {description: 'contact methods',            path: '/about-me/voter/contact-choice'},
+  {description: 'contact details',            path: '/about-me/voter/email-phone'},
+  {description: 'success visit',              path: '/about-me/success-visit'}
+
 ];
 
 const Home = () => {
