--- client/presentations/voter/email-phone-form.jsx
@@ -1,16 +1,0 @@
-'use strict';
-
-import React                    from 'react';
-import HomeLink                 from '../home-link.jsx';
-
-const EmailPhone = (props) => {
-  return (
-    <div>
-      <HomeLink />
-        <form className='email-phone-form'>
-        </form>
-    </div>
-  );
-};
-
-export default EmailPhone;
